fix: return stream startup errors before SSE
