+++ sdk/api/handlers/handlers_stream.go
@@ -1,0 +1,194 @@
+package handlers
+
+import (
+	"context"
+	"errors"
+	"net/http"
+
+	coreexecutor "github.com/router-for-me/CLIProxyAPI/v6/sdk/cliproxy/executor"
+	sdktranslator "github.com/router-for-me/CLIProxyAPI/v6/sdk/translator"
+)
+
+// ExecuteStreamWithAuthManager executes a streaming request via the core auth manager.
+// This path is the only supported execution route.
+// The returned http.Header carries upstream response headers captured before streaming begins.
+func (h *BaseAPIHandler) ExecuteStreamWithAuthManager(ctx context.Context, handlerType, modelName string, rawJSON []byte, alt string) (<-chan []byte, http.Header, <-chan *StreamErrorMessage) {
+	dataChan, upstreamHeaders, errChan, startErr := h.StartStreamWithAuthManager(ctx, handlerType, modelName, rawJSON, alt)
+	if startErr == nil {
+		return dataChan, upstreamHeaders, errChan
+	}
+	legacyErrChan := make(chan *StreamErrorMessage, 1)
+	legacyErrChan <- startErr
+	close(legacyErrChan)
+	return nil, nil, legacyErrChan
+}
+
+// StartStreamWithAuthManager starts a streaming request via the core auth manager.
+// Errors returned directly from this method happen before any downstream stream
+// bytes are written, so HTTP handlers can still return a normal JSON error.
+func (h *BaseAPIHandler) StartStreamWithAuthManager(ctx context.Context, handlerType, modelName string, rawJSON []byte, alt string) (<-chan []byte, http.Header, <-chan *StreamErrorMessage, *StreamErrorMessage) {
+	providers, normalizedModel, errMsg := h.getRequestDetails(ctx, modelName)
+	if errMsg != nil {
+		return nil, nil, nil, errMsg
+	}
+	reqMeta := requestExecutionMetadata(ctx)
+	reqMeta[coreexecutor.RequestedModelMetadataKey] = normalizedModel
+	payload := rawJSON
+	if len(payload) == 0 {
+		payload = nil
+	}
+	req := coreexecutor.Request{
+		Model:   normalizedModel,
+		Payload: payload,
+	}
+	opts := coreexecutor.Options{
+		Stream:          true,
+		Alt:             alt,
+		OriginalRequest: rawJSON,
+		SourceFormat:    sdktranslator.FromString(handlerType),
+	}
+	opts.Metadata = reqMeta
+	streamResult, err := h.AuthManager.ExecuteStream(ctx, providers, req, opts)
+	if err != nil {
+		status := http.StatusInternalServerError
+		if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+			status = 499 // Client Closed Request
+		} else if se, ok := err.(interface{ StatusCode() int }); ok && se != nil {
+			if code := se.StatusCode(); code > 0 {
+				status = code
+			}
+		}
+		var addon http.Header
+		if he, ok := err.(interface{ Headers() http.Header }); ok && he != nil {
+			if hdr := he.Headers(); hdr != nil {
+				addon = hdr.Clone()
+			}
+		}
+		return nil, nil, nil, &StreamErrorMessage{StatusCode: status, Error: err, Addon: addon}
+	}
+	passthroughHeadersEnabled := PassthroughHeadersEnabled(h.Cfg)
+	// Capture upstream headers from the initial connection synchronously before the goroutine starts.
+	// Keep a mutable map so bootstrap retries can replace it before first payload is sent.
+	var upstreamHeaders http.Header
+	if passthroughHeadersEnabled {
+		upstreamHeaders = cloneHeader(FilterUpstreamHeaders(streamResult.Headers))
+		if upstreamHeaders == nil {
+			upstreamHeaders = make(http.Header)
+		}
+	}
+	chunks := streamResult.Chunks
+	dataChan := make(chan []byte)
+	errChan := make(chan *StreamErrorMessage, 1)
+	go func() {
+		defer close(dataChan)
+		defer close(errChan)
+		sentPayload := false
+		bootstrapRetries := 0
+		maxBootstrapRetries := StreamingBootstrapRetries(h.Cfg)
+
+		sendErr := func(msg *StreamErrorMessage) bool {
+			if ctx == nil {
+				errChan <- msg
+				return true
+			}
+			select {
+			case <-ctx.Done():
+				return false
+			case errChan <- msg:
+				return true
+			}
+		}
+
+		sendData := func(chunk []byte) bool {
+			if ctx == nil {
+				dataChan <- chunk
+				return true
+			}
+			select {
+			case <-ctx.Done():
+				return false
+			case dataChan <- chunk:
+				return true
+			}
+		}
+
+		bootstrapEligible := func(err error) bool {
+			status := statusFromError(err)
+			if status == 0 {
+				return true
+			}
+			switch status {
+			case http.StatusUnauthorized, http.StatusForbidden, http.StatusPaymentRequired,
+				http.StatusRequestTimeout, http.StatusTooManyRequests:
+				return true
+			default:
+				return status >= http.StatusInternalServerError
+			}
+		}
+
+	outer:
+		for {
+			for {
+				var chunk coreexecutor.StreamChunk
+				var ok bool
+				if ctx != nil {
+					select {
+					case <-ctx.Done():
+						return
+					case chunk, ok = <-chunks:
+					}
+				} else {
+					chunk, ok = <-chunks
+				}
+				if !ok {
+					return
+				}
+				if chunk.Err != nil {
+					streamErr := chunk.Err
+					// Safe bootstrap recovery: if the upstream fails before any payload bytes are sent,
+					// retry a few times (to allow auth rotation / transient recovery) and then attempt model fallback.
+					if !sentPayload && bootstrapRetries < maxBootstrapRetries && bootstrapEligible(streamErr) {
+						bootstrapRetries++
+						retryResult, retryErr := h.AuthManager.ExecuteStream(ctx, providers, req, opts)
+						if retryErr == nil {
+							if passthroughHeadersEnabled {
+								replaceHeader(upstreamHeaders, FilterUpstreamHeaders(retryResult.Headers))
+							}
+							chunks = retryResult.Chunks
+							continue outer
+						}
+						streamErr = retryErr
+					}
+
+					status := http.StatusInternalServerError
+					if se, ok := streamErr.(interface{ StatusCode() int }); ok && se != nil {
+						if code := se.StatusCode(); code > 0 {
+							status = code
+						}
+					}
+					var addon http.Header
+					if he, ok := streamErr.(interface{ Headers() http.Header }); ok && he != nil {
+						if hdr := he.Headers(); hdr != nil {
+							addon = hdr.Clone()
+						}
+					}
+					_ = sendErr(&StreamErrorMessage{StatusCode: status, Error: streamErr, Addon: addon})
+					return
+				}
+				if len(chunk.Payload) > 0 {
+					if handlerType == "openai-response" {
+						if err := validateSSEDataJSON(chunk.Payload); err != nil {
+							_ = sendErr(&StreamErrorMessage{StatusCode: http.StatusBadGateway, Error: err})
+							return
+						}
+					}
+					sentPayload = true
+					if okSendData := sendData(cloneBytes(chunk.Payload)); !okSendData {
+						return
+					}
+				}
+			}
+		}
+	}()
+	return dataChan, upstreamHeaders, errChan, nil
+}
