+++ sdk/api/handlers/openai/openai_responses_handlers_stream_error_test.go
@@ -44,6 +44,35 @@ func (e *successfulResponsesStreamExecutor) HttpRequest(context.Context, *coreau
 	return nil, errors.New("not implemented")
 }
 
+type startFailureResponsesStreamExecutor struct{}
+
+func (e *startFailureResponsesStreamExecutor) Identifier() string { return "codex" }
+
+func (e *startFailureResponsesStreamExecutor) Execute(context.Context, *coreauth.Auth, coreexecutor.Request, coreexecutor.Options) (coreexecutor.Response, error) {
+	return coreexecutor.Response{}, errors.New("not implemented")
+}
+
+func (e *startFailureResponsesStreamExecutor) ExecuteStream(context.Context, *coreauth.Auth, coreexecutor.Request, coreexecutor.Options) (*coreexecutor.StreamResult, error) {
+	return nil, &coreauth.Error{
+		Code:       "no_auth_available",
+		Message:    "no auth available",
+		Retryable:  true,
+		HTTPStatus: http.StatusServiceUnavailable,
+	}
+}
+
+func (e *startFailureResponsesStreamExecutor) Refresh(ctx context.Context, auth *coreauth.Auth) (*coreauth.Auth, error) {
+	return auth, nil
+}
+
+func (e *startFailureResponsesStreamExecutor) CountTokens(context.Context, *coreauth.Auth, coreexecutor.Request, coreexecutor.Options) (coreexecutor.Response, error) {
+	return coreexecutor.Response{}, errors.New("not implemented")
+}
+
+func (e *startFailureResponsesStreamExecutor) HttpRequest(context.Context, *coreauth.Auth, *http.Request) (*http.Response, error) {
+	return nil, errors.New("not implemented")
+}
+
 func TestForwardResponsesStreamTerminalErrorUsesResponsesErrorChunk(t *testing.T) {
 	gin.SetMode(gin.TestMode)
 	base := handlers.NewBaseAPIHandlers(&sdkconfig.SDKConfig{}, nil)
@@ -108,5 +137,45 @@ func TestOpenAIResponsesStreamingSuccessOverridesPreset404Status(t *testing.T) {
 	}
 	if !strings.Contains(recorder.Body.String(), `"type":"response.completed"`) {
 		t.Fatalf("expected successful SSE body, got %q", recorder.Body.String())
+	}
+}
+
+func TestOpenAIResponsesStreamingStartFailureReturnsJSONErrorBeforeSSE(t *testing.T) {
+	gin.SetMode(gin.TestMode)
+
+	executor := &startFailureResponsesStreamExecutor{}
+	manager := coreauth.NewManager(nil, nil, nil)
+	manager.RegisterExecutor(executor)
+
+	auth := &coreauth.Auth{ID: "auth-start-failure", Provider: executor.Identifier(), Status: coreauth.StatusActive}
+	if _, err := manager.Register(context.Background(), auth); err != nil {
+		t.Fatalf("Register auth: %v", err)
+	}
+	registry.GetGlobalRegistry().RegisterClient(auth.ID, auth.Provider, []*registry.ModelInfo{{ID: "test-model"}})
+	t.Cleanup(func() {
+		registry.GetGlobalRegistry().UnregisterClient(auth.ID)
+	})
+
+	base := handlers.NewBaseAPIHandlers(&sdkconfig.SDKConfig{}, manager)
+	h := NewOpenAIResponsesAPIHandler(base)
+
+	recorder := httptest.NewRecorder()
+	c, _ := gin.CreateTestContext(recorder)
+	c.Request = httptest.NewRequest(http.MethodPost, "/v1/responses", strings.NewReader(`{"model":"test-model","stream":true}`))
+	c.Request.Header.Set("Content-Type", "application/json")
+
+	h.Responses(c)
+
+	if recorder.Code != http.StatusServiceUnavailable {
+		t.Fatalf("status = %d, want %d; body=%s", recorder.Code, http.StatusServiceUnavailable, recorder.Body.String())
+	}
+	if got := recorder.Header().Get("Content-Type"); !strings.Contains(got, "application/json") {
+		t.Fatalf("Content-Type = %q, want application/json", got)
+	}
+	if strings.Contains(recorder.Body.String(), "event: error") {
+		t.Fatalf("expected JSON error before SSE starts, got %q", recorder.Body.String())
+	}
+	if !strings.Contains(recorder.Body.String(), "no auth available") {
+		t.Fatalf("expected upstream start error body, got %q", recorder.Body.String())
 	}
 }
